requirements.txt
argh==0.26.2
distro==1.5.0
numpy==1.20.2
pandas==1.2.4
pathtools==0.1.2
Pillow==8.2.0
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
reportlab==3.5.67
six==1.15.0
tabula-py==2.2.0
watchdog==2.0.3
